apps/api/requirements.txt
alembic==1.14.1
fastapi==0.115.8
httpx==0.28.1
orjson==3.10.15
psycopg[binary]==3.2.5
PyJWT==2.10.1
pydantic==2.10.6
pydantic-settings==2.7.1
pytest==8.3.4
pytest-asyncio==0.25.3
python-multipart==0.0.20
SQLAlchemy==2.0.38
uvicorn[standard]==0.34.0
